Migrate `multiprocessing_patches/__init__.py`
line_profiler/_child_process_profiling/multiprocessing_patches/
    Split the contents of `__init__.py` into separate submodules:

    poller.py::Poller
        Migrated from `__init__.py::_Poller`
    _queue.py::PutWrapper
        Migrated from `__init__.py::_QueuePutWrapper`
    config.py::MPConfig
        Migrated from `__init__.py::MPConfig`
    _infrastructure.py
        Patch
            - Migrated from `__init__.py::_Patch`
            - Added attribute `.priority` to the interface
        SingleModulePatch
            Migrated from `__init__.py::Patch`
        Registry
            - Migrated from `__init__.py::_PatchRegistry`
            - Added class method `.from_entry_point()` for constructing
              an instance from patches loaded from an entry point
    _mandatory_patches.py::{RebootForkserver,ResourceTracker,RunPy}Patch
        - Migrated from the eponymous classes in `__init__.py`
        - Now declared as `line_profiler._multiproc_patches` entry
          points
    _profiling_patches.py
        POOL_PATCH, PROCESS_PATCH
            `SingleModulePatch` objects that are now exposed as
            `line_profiler._multiproc_patches` entry points
        wrap_worker[_{write_on_exit,write_per_task}]()
            Migrated from `__init__.py::wrap_worker_pool*()`
        wrap_bootstrap(), wrap_terminate()
            Migrated from the eponymous functions in `__init__.py`
    _optional_patches.py
        CHILD_PIDS_PATCH, LOGGING_PATCH
            `SingleModulePatch` objects that are now exposed as
            `line_profiler._multiproc_patches` entry points
        wrap_handle_results(), wrap_process()
            Migrated from the eponymous functions in `__init__.py`
        wrap_worker()
            Migrated from `__init__.py::wrap_worker_pid()`
    __init__.py::__all__
        Now including the following:
        - `MPConfig` (= `config.py::MPConfig`)
        - `Poller` (= `poller.py::Poller`)
        - `Registry` (= `_infrastructure.py::Registry`)
        - `Timeout` (= `poller.py::Poller.Timeout`)

tests/test_child_procs.py
    Updated imports from
    `line_profiler._child_process_profiling.multiprocessing_patches`

setup.py
    Added the aforementioned `line_profiler._multiproc_patches` entry
    points
