+++ line_profiler/_child_process_profiling/multiprocessing_patches/_mandatory_patches.py
@@ -1,0 +1,168 @@
+from __future__ import annotations
+
+import multiprocessing
+from collections.abc import Callable
+from functools import partial
+from types import MappingProxyType as mappingproxy, MethodType
+from typing import ClassVar, cast
+
+try:
+    from multiprocessing import spawn
+except ImportError:
+    _CAN_USE_SPAWN = False
+else:
+    _CAN_USE_SPAWN = True
+try:
+    from multiprocessing import forkserver
+except ImportError:
+    _CAN_USE_FORKSERVER = False
+else:
+    _CAN_USE_FORKSERVER = (
+        'forkserver' in multiprocessing.get_all_start_methods()
+    )
+try:
+    from multiprocessing import resource_tracker
+except ImportError:
+    _CAN_USE_RESOURCE_TRACKER = False
+else:
+    _CAN_USE_RESOURCE_TRACKER = True
+
+from ..cache import LineProfilingCache
+from ..runpy_patches import create_runpy_wrapper
+
+
+__all__ = ('RebootForkserverPatch', 'ResourceTrackerPatch', 'RunpyPatch')
+
+
+class RebootForkserverPatch:
+    """
+    Reboot the process backing the global
+    :py:class:`multiprocessing.forkserver.ForkServer` instance:
+
+    - When the patch is applied, so as to ensure that child processes
+      forked therefrom actually receives the active patches; and
+
+    - When the session cache is cleaned up, so that child processes
+      forked therefrom is no longer polluted by the patches.
+
+    Note:
+        This uses
+        :py:method:`multiprocessing.forkserver.ForkServer._stop()` which
+        is private API, but it's the same hack used in Python's own test
+        suite -- see the comment to said method.
+    """
+    summary: ClassVar[mappingproxy[str, frozenset[str]]] = mappingproxy({})
+    priority: ClassVar[float | None] = None
+
+    @classmethod
+    def apply(cls, cache: LineProfilingCache, **_) -> None:
+        if not _CAN_USE_FORKSERVER:
+            return
+        cls.reboot()
+        cache.add_cleanup(cls.reboot)
+
+    @staticmethod
+    def reboot() -> None:
+        # Appease the type-checker since `._stop()` is not public API
+        stop = getattr(forkserver._forkserver, '_stop', None)
+        assert callable(stop)
+        stop()
+
+
+class ResourceTrackerPatch:
+    """
+    Patch :py:mod:`multiprocessing.resource_tracker` so that
+    :py:func:`multiprocessing.resource_tracker.ensure_running` and the
+    eponymous method of
+    :py:class:`multiprocessing.resource_tracker.ResourceTracker` report
+    the resource-tracker server PIDs to the session cache.
+
+    Note:
+        The ``ResourceTracker`` server process is spawned when the first
+        :py:mod:`multiprocessing` child process is created via the
+        ``spawn`` or ``forkserver`` start methods. While this server
+        process does not meaningfully contribute to the profiling result
+        either way, since it can be created with profiling set up, its
+        longevity means that :py:meth:`.LineProfilingCache.gather_stats`
+        often catches empty .lprof files which it has occupied but not
+        written to.
+
+        To reduce noise while keeping the empty-file warning for other
+        output files, we report the PIDs used by the server to the
+        session cache so that they can be ignored if necessary.
+    """
+    if _CAN_USE_RESOURCE_TRACKER:
+        summary: ClassVar[mappingproxy[str, frozenset[str]]] = mappingproxy({
+            'multiprocessing.resource_tracker':
+            frozenset({'ensure_running'}),
+            'multiprocessing.resource_tracker.ResourceTracker':
+            frozenset({'ensure_running'}),
+        })
+    else:
+        summary = mappingproxy({})
+    priority: ClassVar[float | None] = None
+
+    @staticmethod
+    @LineProfilingCache._method_wrapper
+    def wrap_ensure_running(
+        cache: LineProfilingCache,
+        vanilla_impl: Callable[['resource_tracker.ResourceTracker'], None],
+        self: 'resource_tracker.ResourceTracker',
+    ) -> None:
+        """
+        Wrap around :py:meth:`multiprocessing.resource_tracker\
+.ResourceTracker.ensure_running`
+        so that the session cache can keep track of the PIDs used by the
+        resource-tracer server.
+        """
+        maybe_pids: set[int | None] = {getattr(self, '_pid', None)}
+        try:
+            vanilla_impl(self)
+        finally:
+            maybe_pids.add(getattr(self, '_pid', None))
+            pids = cast(set[int], maybe_pids - {None})
+            if pids:
+                cache._warn_possible_lack_of_stats(pids)
+
+    @classmethod
+    def apply(
+        cls, cache: LineProfilingCache, *, cleanup: bool = True, **_,
+    ) -> list[str]:
+        if _CAN_USE_RESOURCE_TRACKER:
+            patch = partial(cache.patch, cleanup=cleanup)
+            # Patch the method on the class
+            method = resource_tracker.ResourceTracker.ensure_running
+            method = cls.wrap_ensure_running(method)
+            patch(resource_tracker.ResourceTracker, 'ensure_running', method)
+            # Patch the preexisting bound method on the module
+            instance = resource_tracker._resource_tracker
+            bound_method = MethodType(method, instance)
+            patch(resource_tracker, 'ensure_running', bound_method)
+        return list(cls.summary)
+
+
+class RunpyPatch:
+    """
+    Patch the copy of :py:mod:`runpy` in the
+    :py:mod:`multiprocessing.spawn` namespace so that subprocesses can
+    perform rewrite-based profiling as with
+    :py:func:`line_profiler.autoprofile.autoprofile.run`.
+
+    See also:
+        :py:mod:`line_profiler._child_process_profiling.runpy_patches`
+    """
+    summary: ClassVar[mappingproxy[str, frozenset[str]]]
+    if _CAN_USE_SPAWN and hasattr(spawn, 'runpy'):
+        summary = mappingproxy({'multiprocessing.spawn': frozenset({'runpy'})})
+    else:
+        summary = mappingproxy({})
+    priority: ClassVar[float | None] = None
+
+    @classmethod
+    def apply(
+        cls, cache: LineProfilingCache, *, cleanup: bool = True, **_,
+    ) -> list[str]:
+        if cls.summary:
+            patch = partial(cache.patch, cleanup=cleanup)
+            patch(spawn, 'runpy', create_runpy_wrapper(cache))
+        return list(cls.summary)
